backend/pyproject.toml
[project]
name = "sovereign-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.2",
    "openai>=2.29.0",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "telethon>=1.42.0",
    "uvicorn[standard]>=0.42.0",
]
